requirements.txt
speech_recognition==3.10.0
pyttsx3==2.90
openai==1.35.0
pyaudio==0.2.11
python-dotenv==1.0.0
requests==2.31.0
